Update tests for Java 6.
Some of the XML related tests started failing in Java 6 due to a
difference in the way a org.w3c.dom.Document is written into XML
between Java 5 and Java 6. This change fixes the tests by changing the
test methods to actually parse the golden files and compare the
information in those files to the expected information in the files.